Need help with Game class (read description please).
I am trying to get graphics to display properly in the window when the
Game class is running the game.  We need to be able to paint the
background image and other images as well.  I will be able to implement
hit boxes, entity movements, etc., but I really need some help getting
images to display.  Only thing I've been able to do so far is get the
screen background colored one of the default colors.