Major update:  Large Alien ships 90% functional  (just run the game and look at 'em.  heh-heh).
Large Alien ships now spawn randomly and have several aspects:
1)  They move on their own,
2) their movement is random, as if piloted by someone,
3)  they stay within bounds (y is bounded, x is wrapped),
4)  they fire on their own at seemingly random intervals

NOTE:  all that is left for LargeAlien is to implement hit boxes (which
applies to everything), as well as laser/shell life span.  I will take
care of these (possibly tomorrow night)

My next update will include fixes such that lasers and shells are
deleted after 2 seconds or so, giving enough time to cross the screen or
wrap a couple of time, but not such that they stay long enough to lag
the game with too many objects being managed.