Major Update:  Program now runs as basic game.
Features implemented:
1)  Entity health:  all entities now have health levels that are based
on what the user selects from the main menu,
2)  Planet health:  the game now keeps track of planet health and
decreases it every time a meteor lands,
3)  Shell/Laser lifespan:  tank shells and alien lasers are deleted
after appearing on the screen for a certain amount of time (this is an
efficiency update),
4)  Hit boxes:  shell strikes against all aliens and meteors are now
recognized and processed appropriately and meteor/laser/small alien hits
on the tank are functional as well, and
5)  Entity removal:  Meteors and all aliens are removed once the user
hits them enough to bring their health down to zero or they "collide"
with the planet.

PLEASE NOTE:  User health is tracked appropriately.  Once a game over
screen is set up, we can use tank and planet health to determine if game
over screen needs to be shown.

BUG:  I didn't recognize this would happen at first, but the way I coded
it causes the difficulty to default to hard when no difficulty is
selected from the main menu.  Is this okay or should I fix it to have
easy or normal as default?  Please let me know Tuesday.