Working sound implemented  :D
Sound implementation has been moved to another class and a workaround
has been obtained.  Each sound file is read in and played each time the
sound is needed.

Unless I or someone else figures out how to read in a sound only once
and get it to play multiple times, this current implementation will
work.  Additionally, it does not seem to lag the game or computer at all
or cause any errors in the game.

Note:  each sound file itself I have made and plan to make the rest.  If
it sounds too cheesy or something, we're going to have to find a sound
pack with a free license agreement attached to it.